# RFC: clean up `map` and `collect` with iterator traits