--- a/test/sysinfo.jl
+++ b/test/sysinfo.jl
@@ -12,6 +12,10 @@ Base.Sys.loadavg()
 @test_throws ArgumentError ENV["okname"] = "bad\0val"
 @test_throws ArgumentError Sys.set_process_title("bad\0title")
 
+withenv("bad"=>"dog") do
+    @test_throws ArgumentError ENV["bad\0cat"]
+end
+
 # issue #11170
 withenv("TEST"=>"nonempty") do
     @test ENV["TEST"] == "nonempty"
@@ -25,3 +29,9 @@ let c = collect(ENV)
     @test length(ENV) == length(c)
     @test isempty(ENV) || first(ENV) in c
 end
+
+# test for non-existent keys
+key = randstring(25)
+@test !haskey(ENV,key)
+@test_throws KeyError ENV[key]
+@test get(ENV,key,"default") == "default"