This is a proposal for cleaning up our iteration, mapping, and collecting. Between `copy[!]`, `collect`, `map`, and comprehensions we had subtly different behaviors for similar operations. We also had some repeated code and inefficient generic fallbacks.

This change replaces all that with `collect` of various iterators. There are four core algorithms for this, based on whether an iterator has a known length/shape and whether it has a known element type. To select the right algorithm, I added some HolyTraits for iterators. I seem to recall iterator traits being discussed before, but I can't find a reference.

The biggest immediate benefit is that generic `map` is faster and better-typed.
Before:

```
julia> @time map(tuple, 11:2000000, enumerate(11:2000000));
  2.745751 seconds (18.00 M allocations: 505.257 MB, 29.04% gc time)
```

giving a `Vector{Any}`.
After:

```
julia> @time map(tuple, 11:2000000, enumerate(11:2000000));
  0.570696 seconds (6.00 M allocations: 198.356 MB, 11.51% gc time)
```

giving a `Vector{Tuple{Int64,Tuple{Int64,Int64}}}`.
This is achieved with the following glorious definition(s):

```
spread(f) = (args)->f(args...)
map(f, iters...) = collect(Generator(spread(f),zip(iters...)))
```

This definition is *almost* as fast as existing array functions:

```
julia> A = rand(1000,1000);

julia> @time A+A;
  0.001964 seconds (9 allocations: 7.630 MB)

julia> @time map(+,A,A);
  0.004792 seconds (11 allocations: 7.630 MB)
```

Hopefully this gap can be closed.

A major issue here is iterator shape preservation. Some iterators just wrap others and so should obviously preserve their shape, like `Generator`. Others are less obvious. Since zipped iteration is so common when working with arrays, I made `zip` shape-preserving. `product` is currently 1d only, but could for example concatenate the shapes of its component iterators.