@tbreloff 2016-02-18
+1. Does this mean we could have a correctly typed unzip as well?

On Wednesday, February 17, 2016, Jeff Bezanson <notifications@github.com>
wrote:

> This is a proposal for cleaning up our iteration, mapping, and collecting.
> Between copy[!], collect, map, and comprehensions we had subtly different
> behaviors for similar operations. We also had some repeated code and
> inefficient generic fallbacks.
>
> This change replaces all that with collect of various iterators. There
> are four core algorithms for this, based on whether an iterator has a known
> length/shape and whether it has a known element type. To select the right
> algorithm, I added some HolyTraits for iterators. I seem to recall iterator
> traits being discussed before, but I can't find a reference.
>
> The biggest immediate benefit is that generic map is faster and
> better-typed.
> Before:
>
> julia> @time map(tuple, 11:2000000, enumerate(11:2000000));
>   2.745751 seconds (18.00 M allocations: 505.257 MB, 29.04% gc time)
>
> giving a Vector{Any}.
> After:
>
> julia> @time map(tuple, 11:2000000, enumerate(11:2000000));
>   0.570696 seconds (6.00 M allocations: 198.356 MB, 11.51% gc time)
>
> giving a Vector{Tuple{Int64,Tuple{Int64,Int64}}}.
> This is achieved with the following glorious definition(s):
>
> spread(f) = (args)->f(args...)
> map(f, iters...) = collect(Generator(spread(f),zip(iters...)))
>
> This definition is *almost* as fast as existing array functions:
>
> julia> A = rand(1000,1000);
>
> julia> @time A+A;
>   0.001964 seconds (9 allocations: 7.630 MB)
>
> julia> @time map(+,A,A);
>   0.004792 seconds (11 allocations: 7.630 MB)
>
> Hopefully this gap can be closed.
>
> A major issue here is iterator shape preservation. Some iterators just
> wrap others and so should obviously preserve their shape, like Generator.
> Others are less obvious. Since zipped iteration is so common when working
> with arrays, I made zip shape-preserving. product is currently 1d only,
> but could for example concatenate the shapes of its component iterators.
> ------------------------------
> You can view, comment on, or merge this pull request online at:
>
>   https://github.com/JuliaLang/julia/pull/15123
> Commit Summary
>
>    - make `map` and `collect` more general and uniform by adding iterator
>    traits
>
> File Changes
>
>    - *M* base/abstractarray.jl
>    <https://github.com/JuliaLang/julia/pull/15123/files#diff-0> (150)
>    - *M* base/array.jl
>    <https://github.com/JuliaLang/julia/pull/15123/files#diff-1> (99)
>    - *M* base/dict.jl
>    <https://github.com/JuliaLang/julia/pull/15123/files#diff-2> (3)
>    - *M* base/generator.jl
>    <https://github.com/JuliaLang/julia/pull/15123/files#diff-3> (39)
>    - *M* base/inference.jl
>    <https://github.com/JuliaLang/julia/pull/15123/files#diff-4> (2)
>    - *M* base/iterator.jl
>    <https://github.com/JuliaLang/julia/pull/15123/files#diff-5> (46)
>    - *M* test/abstractarray.jl
>    <https://github.com/JuliaLang/julia/pull/15123/files#diff-6> (14)
>    - *M* test/arrayops.jl
>    <https://github.com/JuliaLang/julia/pull/15123/files#diff-7> (2)
>    - *M* test/datafmt.jl
>    <https://github.com/JuliaLang/julia/pull/15123/files#diff-8> (2)
>    - *M* test/unicode/utf8proc.jl
>    <https://github.com/JuliaLang/julia/pull/15123/files#diff-9> (2)
>
> Patch Links:
>
>    - https://github.com/JuliaLang/julia/pull/15123.patch
>    - https://github.com/JuliaLang/julia/pull/15123.diff
>
> —
> Reply to this email directly or view it on GitHub
> <https://github.com/JuliaLang/julia/pull/15123>.
>